+++ src/main/java/Domain/ProjectState.java
@@ -4,7 +4,7 @@ import Common.DTO.BitDTO;
 import Common.DTO.ProjectStateDTO;
 import Common.DTO.VertexDTO;
 import Common.Exceptions.InvalidBitException;
-import Domain.Interfaces.IMemorizer;
+import Common.Interfaces.IMemorizer;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -21,7 +21,7 @@ class ProjectState {
     private final Bit[] bitList;
     private PanelCNC panel;
     private static final VertexDTO defaultPanelDimension = new VertexDTO(1219.2, 914.4, 0); // dimension in mm
-    private static final float defaultPanelDepth = 1.0f; // depth in mm
+    private static final double defaultPanelDepth = 1.0f; // depth in mm
     private final IMemorizer memorizer;
     private Map<Integer, BitDTO> configuredBits;
